itemselectionmodel: Move BlockSelectionModel to utils
